add category component and dynamic generic query method

+++ ShoeStore/Services/ProductService.cs
@@ -1,4 +1,6 @@
-﻿using ShoeStore.Entities;
+﻿using AutoMapper;
+using ShoeStore.DTO;
+using ShoeStore.Entities;
 using ShoeStore.Repositories.Implementation;
 using ShoeStore.Services.Implementation;
 
@@ -7,37 +9,55 @@ namespace ShoeStore.Services
     public class ProductService : IProductService
     {
         private readonly IRepository<Product> _productRepository;
+        private readonly IMapper _mapper;
 
-        public ProductService(IRepository<Product> productRepo) 
+        public ProductService(IRepository<Product> productRepo, IMapper mapper) 
         { 
             this._productRepository = productRepo;
+            this._mapper = mapper;
         }
 
-        public async Task AddProductAsync(Product product)
+        public async Task AddProductAsync(ProductDTO dto)
         {
-            await _productRepository.AddAsync(product);
+            await _productRepository.AddAsync(_mapper.Map<Product>(dto));
         }
 
-        public async Task DeleteProductAsync(int id)
+        public async Task DeleteProductAsync(int id)    
         {
+            if(_productRepository.GetByIdAsync(id) == null)
+            {
+                throw new Exception("Product not found");
+            }
             await _productRepository.DeleteAsync(id);
         }
 
-        public async Task<IEnumerable<Product>> GetAllProductsAsync()
+        public async Task<IEnumerable<ProductDTO>> GetAllProductsAsync()
         {
-            return await _productRepository.GetAllAsync();
+            IEnumerable<Product> products = await _productRepository.GetAllAsync();
+            return _mapper.Map<IEnumerable<ProductDTO>>(products);
         }
 
-        public async Task<Product> GetProductByIdAsync(int id)
+        public async Task<ProductDTO> GetProductByIdAsync(int id)
         {
-            return await _productRepository.GetByIdAsync(id);
+            Product product = await _productRepository.GetByIdAsync(id);
+            return _mapper.Map<ProductDTO>(product);
+
         }
 
-        public async Task UpdateProductAsync(Product product)
+        public async Task UpdateProductAsync(ProductDTO dto)
         {
-            await _productRepository.UpdateAsync(product);
+            if(GetProductByIdAsync(dto.Id) == null)
+            {
+                throw new Exception("Product not found");
+            }
+            await _productRepository.UpdateAsync(_mapper.Map<Product>(dto));
         }
 
-        
+        public async Task<List<ProductDTO>> GetProductListByCategory(int categoryId)
+        {
+            List<Product> products = await _productRepository.FindByConditionAsync(p => p.CategoryId == categoryId , p => p.Category);
+            return _mapper.Map<List<ProductDTO>>(products);
+        }
+
     }
 }
